admin_portal/requirements.txt
altair==5.1.2
attrs==23.1.0
blinker==1.7.0
cachetools==5.3.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
gitdb==4.0.11
GitPython==3.1.41
idna==3.7
importlib-metadata==6.8.0
Jinja2==3.1.4
jsonschema==4.20.0
jsonschema-specifications==2023.11.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
numpy==1.26.2
packaging==23.2
pandas==2.1.3
Pillow==10.3.0
protobuf==4.25.1
pyarrow==14.0.1
pydeck==0.8.1b0
Pygments==2.17.2
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
referencing==0.31.0
requests==2.32.2
rich==13.7.0
rpds-py==0.13.1
setuptools==70.0.0
six==1.16.0
smmap==5.0.1
streamlit==1.37.0
tenacity==8.2.3
toml==0.10.2
toolz==0.12.0
tornado==6.4.1
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
urllib3==2.2.2
validators==0.22.0
watchdog==3.0.0
wheel==0.41.2
zipp==3.19.1
